Xavi review

+++ code/Fut7/Application/Season/CRUD/Delete/DeleteSeasonUseCase.php
@@ -3,6 +3,7 @@
 namespace Fut7\Application\Season\CRUD\Delete;
 
 use Fut7\Domain\Contract\Repository\SeasonRepositoryInterface;
+use Fut7\Domain\Entity\Season\Season;
 use Fut7\Domain\Response\Season\DeleteSeasonResponse;
 
 class DeleteSeasonUseCase
@@ -22,8 +23,12 @@ class DeleteSeasonUseCase
         if (null === $uuid) {
             throw new \TypeError("UUID must be a string!!");
         }
-
         $season = $this->seasonRepository->find($uuid);
+        if (null == $season) {
+            throw new \TypeError('Season with uuid '.$uuid.' not found. Skip delete action');
+        }
+
+
         $this->seasonRepository->delete($season);
         return new DeleteSeasonResponse($season);
     }
